stable microbenchmark
