+++ release/microbenchmark/run_microbenchmark.py
@@ -1,6 +1,8 @@
 import json
 import os
 
+import argparse
+
 
 def to_dict_key(key: str):
     for r in [" ", ":", "-"]:
@@ -11,7 +13,18 @@ def to_dict_key(key: str):
 
 
 if __name__ == "__main__":
-    from ray._private.ray_perf import main
+    parser = argparse.ArgumentParser(
+        "--experimental",
+        action="store_true",
+        default=False,
+        help="If passed, run ray.experimental microbenchmarks.",
+    )
+    args = parser.parse_args()
+
+    if args.experimental:
+        from ray._private.ray_experimental_perf import main
+    else:
+        from ray._private.ray_perf import main
 
     results = main() or []
